requirements.txt
slackclient
flask
python-dotenv
opencv-python
Pillow
mysql-connector-python